apps/control-plane/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relay-control-plane"
version = "0.1.0"
description = "Relay Control Plane Lite service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "pydantic[email]>=2.7,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "sqlalchemy>=2.0.25,<3.0",
    "alembic>=1.12,<2.0",
    "passlib[bcrypt]>=1.7.4,<2.0",
    "bcrypt>=3.2,<4.0",
    "PyJWT>=2.8.0,<3.0",
    "cryptography>=41.0.0,<43.0",
    "psycopg[binary]>=3.1,<4.0",
    "uvicorn>=0.23,<1.0",
    "slowapi>=0.1.9,<0.2.0",
    "jinja2>=3.1.0,<4.0",
    "python-multipart>=0.0.9,<0.1.0",
    "authlib>=1.3.0,<2.0",
    "httpx>=0.27.0,<0.28.0",
    "pyotp>=2.9.0,<3.0",
    "qrcode>=7.4,<8.0",
    "pillow>=10.0,<11.0",
    "prometheus-client>=0.20.0,<1.0",
    "python-json-logger>=2.0.0,<3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.1,<9.0",
    "httpx>=0.26,<0.28",
]
dev = [
    "ruff>=0.6.9,<0.7.0",
]

[tool.setuptools]
package-dir = {"" = "app"}

[tool.setuptools.packages.find]
where = ["app"]

[tool.setuptools.package-data]
"app" = [
    "db/migrations/*.py",
    "db/migrations/versions/*.py",
    "db/migrations/script.py.mako",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402", "F841"]  # Allow imports after sys.path manipulation and unused vars in tests
"app/db/migrations/*" = ["F821"]  # Migration downgrade functions reference upgrade variables

[tool.ruff.format]
quote-style = "double"
